Big Update
I am lazy about updating this repository.  Here are a ton of changes to the framework.  Cleaned up code and updated just about every part of this framework.  Still lots of work to do before we can go to beta.  I did add a composer file which should allow us to install this via composer.  Since we are still in the early stages of this project I am not ready to include this in the composer repo so I need to see if I can install directly from git.